how to insert images in HTML
Chapter Summary. Use the HTML <img> element to define an image. Use the HTML src attribute to define the URL of the image. Use the HTML alt attribute to define an alternate text for an image, if it cannot be displayed. Use the HTML width and height attributes to define the size of the image.